requirements.txt
requests
pandas
toml
python-dotenv
boto3